Themes and random recipe

+++ src/utils/users.js
@@ -1,5 +1,5 @@
-import { ALL_USERS, CUR_USER, USER_TO_EDIT } from "../config/localStorage"
-import { HOME, SIGN_IN, MANAGEUSERS, EDITUSER } from "../config/routes"
+import { ALL_USERS, CUR_USER, USER_TO_EDIT,THEME } from "../config/localStorage"
+import { HOME, SIGN_IN, MANAGEUSERS, EDITUSER, USER_SETTINGS } from "../config/routes"
 import { gatherNotifications } from "./notifications"
 import React from 'react';
 import { toast } from 'react-toastify';
@@ -13,9 +13,9 @@ export function checkUserLogin(currentUser, navigate) {
   }
 }
 //Checks is a user is logged in and if it is an admin
-export function checkAdminLogin(currentUser, navigate){
+export function checkAdminLogin(currentUser, navigate) {
   const userData = JSON.parse(currentUser)
-  if(userData.adminlevel === 3){
+  if (userData.adminlevel === 3) {
     checkUserLogin(currentUser, navigate)
   } else {
     navigate(HOME)
@@ -23,11 +23,13 @@ export function checkAdminLogin(currentUser, navigate){
 }
 //Basic login function
 export function logIn(attemptingUser, navigate) {
+  const themeStr = localStorage.getItem(THEME)
+  const theme = JSON.parse(themeStr)
   //Checks if both fields have a value
   if (attemptingUser.username && attemptingUser.password) {
     //Check for user in local storage
     if (validateUser(attemptingUser) === false) {
-      toast("Invalid Username or Password!",{position: "bottom-right"})
+      toast("Invalid Username or Password!", { position: "bottom-right", theme:theme.toast})
     } else {
       localStorage.setItem(CUR_USER, JSON.stringify(attemptingUser))
       navigate(HOME)
@@ -47,6 +49,7 @@ export function validateUser(attemptingUser) {
       attemptingUser.expirylimit = allUserData[i].expirylimit
       attemptingUser.adminlevel = allUserData[i].adminlevel
       attemptingUser.image = allUserData[i].image
+      attemptingUser.friends = allUserData[i].friends
       return true
     }
   }
@@ -67,10 +70,11 @@ export function addUser(userToRegister, navigate) {
         notify: false,
         itemlimit: 10,
         expirylimit: 7,
-        adminlevel: 2
+        adminlevel: 2,
+        friends: []
       }
       //Test newUser against current registered users, then adds to local storage All_USERS               
-      if (userExists(newUser) === false) {
+      if ((userExists(newUser) === false) && (userEmailExists(newUser) === false)) {
         userSave(newUser)
         localStorage.setItem(CUR_USER, JSON.stringify(newUser))
         navigate(HOME)
@@ -91,10 +95,24 @@ export function userExists(userToCheck) {
   }
   return false
 }
+
+export function userEmailExists(userToCheck) {
+  const allUserData = JSON.parse(localStorage.getItem(ALL_USERS))
+  if (allUserData === null || allUserData === "") {
+    return false
+  }
+  for (let i = 0; i < allUserData.length; i++) {
+    if (allUserData[i].email === userToCheck.email) {
+      return true
+    }
+  }
+  return false
+}
+
 //Saves user to local storage, should work without modification
 export function userSave(userToSave) {
   const allUserDataStr = localStorage.getItem(ALL_USERS)
-  userToSave.id ="" + new Date().getTime() + "-" + userToSave.username
+  userToSave.id = "" + new Date().getTime() + "-" + userToSave.username
   let temparr
   if (!(allUserDataStr === null || allUserDataStr.trim() === "")) {
     const allUserData = JSON.parse(allUserDataStr)
@@ -105,21 +123,23 @@ export function userSave(userToSave) {
   localStorage.setItem(ALL_USERS, JSON.stringify(temparr))
 }
 //Save current user to all users in local storage
-export function saveUserSettings(currentUser) {
+export function saveUserSettings(currentUser, navigate) {
   const allUserDataStr = localStorage.getItem(ALL_USERS)
   const allUserData = allUserDataStr ? JSON.parse(allUserDataStr) : []
   const filteredUsers = allUserData.filter(users => !users.id.match(new RegExp('^' + currentUser.id + '$')))
   const newAllUsers = [...filteredUsers, currentUser]
   localStorage.setItem(ALL_USERS, JSON.stringify(newAllUsers))
+  localStorage.setItem(CUR_USER,JSON.stringify(currentUser))
+  navigate(USER_SETTINGS)
 }
 //Changes current users passwords
-export function changeUserPassword(passwords) {
+export function changeUserPassword(passwords,navigate) {
   const currentUser = JSON.parse(localStorage.getItem(CUR_USER))
   if (currentUser.password === passwords.currentpassword) {
     if (passwords.newpassword === passwords.newpasswordcheck) {
       currentUser.password = passwords.newpassword
       localStorage.setItem(CUR_USER, JSON.stringify(currentUser))
-      saveUserSettings(currentUser)
+      saveUserSettings(currentUser,navigate)
       return true
     } else {
       return false
@@ -158,27 +178,29 @@ export function getCurrentUsername() {
 export function displayUsers(navigate) {
   const allUserDataStr = localStorage.getItem(ALL_USERS)
   const allUserData = allUserDataStr ? JSON.parse(allUserDataStr) : []
+  const themeStr = localStorage.getItem(THEME)
+  const theme = JSON.parse(themeStr)
   return (
-    <table className="table table-info table-striped">
+    <table className={theme.table}>
       <tbody>
-        <tr key="header">
-          <th scope="col">Username</th>
-          <th scope="col">Admin Level</th>
-          <th scope="col">Options</th>
+        <tr  key="header">
+          <th  scope="col">Username</th>
+          <th  scope="col">Admin Level</th>
+          <th  scope="col">Options</th>
         </tr>
         {allUserData.map(user => {
           if (user.adminlevel <= 2) {
             return (
               <tr key={user.id}>
-                <td>
+                <td >
                   {user.username}
                 </td>
-                <td>
+                <td >
                   {user.adminlevel}
                 </td>
-                <td>
-                  <button className="btn btn-primary" style = {{marginRight:16}} onClick={() => editUser(user, navigate)}>Edit User</button>
-                  <button className="btn btn-primary" onClick={() => deleteUser(user, navigate)}>Delete User</button>
+                <td >
+                  <button className={theme.button} style={{ marginRight: 16 }} onClick={() => editUser(user, navigate)}>Edit User</button>
+                  <button className={theme.button} onClick={() => deleteUser(user, navigate)}>Delete User</button>
                 </td>
               </tr>
             )
@@ -202,4 +224,59 @@ export function deleteUser(userToDelete, navigate) {
 export function editUser(userToEdit, navigate) {
   localStorage.setItem(USER_TO_EDIT, JSON.stringify(userToEdit))
   navigate(EDITUSER)
+}
+
+export function getUserImage(username) {
+  const allUserDataStr = localStorage.getItem(ALL_USERS)
+  const allUserData = allUserDataStr ? JSON.parse(allUserDataStr) : []
+  const filteredUser = allUserData.filter(users => users.username.match(new RegExp('^' + username + '$')))
+  return filteredUser[0].image
+}
+
+export function getUserIDByEmail(emailToSearch) {
+  const allUserDataStr = localStorage.getItem(ALL_USERS)
+  const allUserData = allUserDataStr ? JSON.parse(allUserDataStr) : []
+  if (allUserData.length > 0) {
+    const filteredUser = allUserData.filter(users => users.email.match(new RegExp('^' + emailToSearch + '$')))
+    if (filteredUser.length > 0) {
+      return filteredUser[0].id
+    } else {
+      return "No User Found"
+    }
+  }
+
+}
+
+export function getUserNameByID(idToSearch){
+  const allUserDataStr = localStorage.getItem(ALL_USERS)
+  const allUserData = allUserDataStr ? JSON.parse(allUserDataStr) : []
+  if (allUserData.length > 0) {
+    const filteredUser = allUserData.filter(users => users.id.match(new RegExp('^' + idToSearch + '$')))
+    if (filteredUser.length > 0) {
+      return filteredUser[0].username
+    } else {
+      return "No User Found"
+    }
+  }
+}
+
+export function addFriend(currentUser, friendToAddID) {
+  const allUserDataStr = localStorage.getItem(ALL_USERS)
+  const allUserData = allUserDataStr ? JSON.parse(allUserDataStr) : []
+  let filteredUsers = allUserData.filter(users => !users.id.match(new RegExp('^' + currentUser.id + '$')))
+  console.log(filteredUsers)
+  console.log(currentUser.id)
+  filteredUsers = filteredUsers.filter(users => !users.id.match(new RegExp('^' + friendToAddID + '$')))
+  console.log(filteredUsers)
+  let friendUserEntry = allUserData.filter(users => users.id.match(new RegExp('^' + friendToAddID + '$')))
+  let friendUser = friendUserEntry[0]
+
+  friendUser.friends.push(currentUser.id)
+  filteredUsers = [...filteredUsers,friendUser]
+  currentUser.friends.push(friendToAddID)
+  filteredUsers = [...filteredUsers,currentUser]
+  
+  localStorage.setItem(CUR_USER,JSON.stringify(currentUser))
+  localStorage.setItem(ALL_USERS,JSON.stringify(filteredUsers))
+  
 }
